Move kernel prover to aztec rpc.
