+++ yarn-project/aztec-rpc/package.json
@@ -40,7 +40,6 @@
     "@aztec/barretenberg.js": "workspace:^",
     "@aztec/circuits.js": "workspace:^",
     "@aztec/foundation": "workspace:^",
-    "@aztec/kernel-prover": "workspace:^",
     "@aztec/noir-contracts": "workspace:^",
     "@aztec/types": "workspace:^",
     "browserify-cipher": "^1.0.1",
