+++ yarn-project/aztec-rpc/src/kernel_prover/kernel_prover.test.ts
