+++ yarn-project/aztec-rpc/src/kernel_prover/proof_creator.ts
