+++ yarn-project/aztec-rpc/src/kernel_prover/proving_data_oracle.ts
