--- yarn-project/kernel-prover/.eslintrc.cjs
@@ -1,1 +1,0 @@
-module.exports = require('@aztec/foundation/eslint-legacy');
