--- yarn-project/kernel-prover/README.md
@@ -1,1 +1,0 @@
-# Kernel Simulator
