--- yarn-project/kernel-prover/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@aztec/kernel-prover",
-  "version": "0.0.0",
-  "type": "module",
-  "exports": "./dest/index.js",
-  "typedoc": {
-    "entryPoint": "./src/index.ts",
-    "displayName": "Kernel Simulator",
-    "tsconfig": "./tsconfig.json"
-  },
-  "scripts": {
-    "prepare": "node ../yarn-project-base/scripts/update_build_manifest.mjs package.json",
-    "prepare:check": "node ../yarn-project-base/scripts/update_build_manifest.mjs package.json --check",
-    "build": "yarn clean && tsc -b",
-    "build:dev": "tsc -b --watch",
-    "clean": "rm -rf ./dest .tsbuildinfo",
-    "formatting": "run -T prettier --check ./src && run -T eslint ./src",
-    "formatting:fix": "run -T prettier -w ./src",
-    "test": "NODE_NO_WARNINGS=1 node --experimental-vm-modules $(yarn bin jest) --passWithNoTests"
-  },
-  "inherits": [
-    "../package.common.json"
-  ],
-  "jest": {
-    "preset": "ts-jest/presets/default-esm",
-    "moduleNameMapper": {
-      "^(\\.{1,2}/.*)\\.js$": "$1"
-    },
-    "testRegex": "./src/.*\\.test\\.ts$",
-    "rootDir": "./src"
-  },
-  "dependencies": {
-    "@aztec/acir-simulator": "workspace:^",
-    "@aztec/circuits.js": "workspace:^",
-    "@aztec/foundation": "workspace:^",
-    "tslib": "^2.4.0"
-  },
-  "devDependencies": {
-    "@jest/globals": "^29.5.0",
-    "@rushstack/eslint-patch": "^1.1.4",
-    "@types/jest": "^29.5.0",
-    "@types/node": "^18.7.23",
-    "jest": "^29.5.0",
-    "ts-jest": "^29.1.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4"
-  },
-  "files": [
-    "dest",
-    "src",
-    "!*.test.*"
-  ],
-  "types": "./dest/index.d.ts"
-}
