--- yarn-project/kernel-prover/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "..",
-  "compilerOptions": {
-    "outDir": "dest",
-    "rootDir": "src",
-    "tsBuildInfoFile": ".tsbuildinfo"
-  },
-  "references": [
-    {
-      "path": "../acir-simulator"
-    },
-    {
-      "path": "../circuits.js"
-    },
-    {
-      "path": "../foundation"
-    }
-  ],
-  "include": ["src"]
-}
